Finally Project Done

+++ app/src/main/java/com/example/weatherapplication/Model/Main.java
@@ -2,7 +2,7 @@ package com.example.weatherapplication.Model;
 
 public class Main {
     private double temp;
-    private double pressure;
+    private float pressure;
     private int humidity;
     private double temp_min;
     private double temp_max;
@@ -18,11 +18,11 @@ public class Main {
         this.temp = temp;
     }
 
-    public double getPressure() {
+    public float getPressure() {
         return pressure;
     }
 
-    public void setPressure(double pressure) {
+    public void setPressure(float pressure) {
         this.pressure = pressure;
     }
